pwm-backlight: Don't enable backlight at probe time for DT
This driver enables backlight devices at probe time. This can be useful,
even essential, for setups that use a simple framebuffer device where no
mechanism exists to control the backlight other than userspace (via the
sysfs interface).

However, when dealing with more complex setups, such as those of recent
ARM SoCs, this can be problematic. Since the backlight is usually setup
separately from the display controller, probe order is nondeterministic
and can cause the backlight to be enabled while the display controller
hasn't been properly initialized yet. This is undesirable because the
display may show garbage.

Furthermore, subsystems such as DRM have advanced functionality to set
the power mode of a panel. In order to allow such setups to power up the
panel at exactly the right moment, a way is needed to prevent the
backlight core from powering the backlight up automatically when it is
registered.

This commit changes the driver to not automatically enable backlight at
probe time when booting using DT. The assumption is that platforms using
DT are modern enough to control the backlight through some higher-level
mechanism such as DRM/KMS.

As a side-effect this change will cause the backlight to remain in what-
ever state firmware might have left it. This helps with transitioning in
a seamless way from firmware to kernel.

Signed-off-by: Thierry Reding <treding@nvidia.com>