swiotlb: Make linux/swiotlb.h standalone includable
This header file uses the enum dma_data_direction and struct page types
without explicitly including the corresponding header files. This makes
it rely on the includer to have included the proper headers before.

To fix this, include linux/dma-direction.h and forward-declare struct
page. The swiotlb_free() function is also annotated __init, therefore
requires linux/init.h to be included as well.

Signed-off-by: Thierry Reding <treding@nvidia.com>
---
Alternatively rather than include linux/init.h for swiotlb_free() it
should be possible to simply remove the annotation in the header since
the implementation already has it and that's where it matters.