PCI: tegra: Track bus -> CPU mapping
Track the offsets of the bus -> CPU mapping for I/O and memory. This is
cosmetic for current Tegra chips because the offset is always 0. But to
properly support legacy use-cases, like VGA, this would be needed so
that PCI bus addresses can be relocated.

While at it, also request the I/O resource both in physical memory and
I/O space to make /proc/iomem consistent, as well as add the I/O region
to the list of host bridge resources.

Signed-off-by: Thierry Reding <treding@nvidia.com>