ALSA: hda/tegra: Make clock and reset names consistent
The example in the device tree binding documentation and the driver code
name the third clock and reset hda2codec_2x, but the documentation names
it hdacodec_2x. On the assumption that this is a typo, fix up the naming
in the device tree binding. Since no board enables the HDA yet, this can
still be changed without breaking backwards-compatibility.

Signed-off-by: Thierry Reding <treding@nvidia.com>