arm64: tegra: Add NVIDIA P2595 base board support
The NVIDIA P2595 base board is used in several reference designs and
provides the connectors to connect the P2530 compute module to the
outside world. It features a USB 3.0 network adapter, a USB 3.0 port,
an HDMI port, a SATA port, an audio codec, a microSD card slot and a
display connector, among others.

Signed-off-by: Thierry Reding <treding@nvidia.com>