clk: tegra: Implement SOR1 clock on Tegra210
SOR1 clock uses a two-stage mux to select a parent. This is difficult to
represent in software, especially if drivers need to select one of the
parents during the programming of the SOR module.

To model the two-stage mux using a single level of parents, use a parent
and mask/value mapping.

Signed-off-by: Thierry Reding <treding@nvidia.com>