drm/tilcdc: panel: Add tilcdc- prefix to driver name
The current driver name, "panel", is very generic. This causes problems
when the driver is enabled on non-TI device tree platforms if the panel
node is called "panel". This causes the driver name match fall-back in
platform_match() to succeed, even if some other driver would precisely
match using the OF device ID table, for example.

Currently there don't seem to be any in-tree users that rely on the
driver name match, so adding the prefix should be safe.

Signed-off-by: Thierry Reding <treding@nvidia.com>