firmware: tegra: Add BPMP support
The Boot and Power Management Processor (BPMP) is a co-processor found
on Tegra SoCs. It is designed to handle the early stages of the boot
process and offload power management tasks (such as clocks, resets,
powergates, ...) as well as system control services.

Compared to the ARM SCPI, the services provided by BPMP are message-
based rather than method-based. The BPMP firmware driver provides the
services to transmit data to and receive data from the BPMP. Users can
also register a Message ReQuest (MRQ), for which a service routine will
be run when a corresponding event is received from the firmware.

A set of messages, called the BPMP ABI, are specified for a number of
different services provided by the BPMP (such as clocks or resets).

Based on work by Sivaram Nair <sivaramn@nvidia.com> and Joseph Lo
<josephl@nvidia.com>.

Signed-off-by: Thierry Reding <treding@nvidia.com>
---
Changes in v4:
- fix race condition between channel setup and interrupt service routine
- correctly check for return value of wait_for_completion_timeout()
- use _size suffix for variables denoting size to avoid confusion
- use correct repyl to BPMP ping challenge
- avoid potential NULL pointer dereference
- use more consistent error handling
- properly clean up on failure
- use defines from bpmp-abi.h
- implement correct timeout

Changes in v3:
- use a message structure for transfers to avoid long function
  prototypes
- restructure driver for easier maintainability
- rename bpmp_abi.h to bpmp-abi.h for consistency