soc/tegra: bpmp: Implement generic PM domains
The BPMP firmware, found on Tegra186 and later, provides an ABI that can
be used to enable and disable power to several power partitions in Tegra
SoCs. The ABI allows for enumeration of the available power partitions,
so the driver can be reused on future generations, provided the BPMP ABI
remains stable.

Based on work by Stefan Kristiansson <stefank@nvidia.com> and Mikko
Perttunen <mperttunen@nvidia.com>.

Signed-off-by: Thierry Reding <treding@nvidia.com>
---
Changes in v2:
- take into account PG_STATE_RUNNING in tegra_bpmp_powergate_is_powered()
- check return value of pm_genpd_init() and clean up on failure
- simplify error unwinding in tegra_bpmp_init_powergates()